pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poiesis"
version = "0.1.0"
description = "Autonomous long-form narrative generation engine"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "PyJWT>=2.7.0",
    "bcrypt>=3.2.0",
    "python-multipart>=0.0.9",
    "cryptography>=42.0.4",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
# 真实语义向量模式所需的重型依赖（PyTorch + transformers）
# 使用 POIESIS_EMBEDDING_MODE=real 时必须安装此扩展
# 安装方式：pip install "poiesis[real-embedding]"
real-embedding = [
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.4.0",
    "types-PyYAML>=6.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
]

[project.scripts]
poiesis = "poiesis.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["poiesis*"]

[tool.setuptools.package-data]
poiesis = ["db/schema.sql"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=poiesis --cov-report=term-missing"
asyncio_mode = "auto"
